fix: fix refactored locations
