+++ src/aind_data_schema/core/instrument.py
@@ -51,7 +51,7 @@ from aind_data_schema.components.devices import (
     Wheel,
     Scanner,
 )
-from aind_data_schema.components.calibrations import Calibration
+from aind_data_schema.components.measurements import Calibration
 
 # Define the mapping of modalities to their required device types
 # The list of list pattern is used to allow for multiple options within a group, so e.g.
