+++ tests/test_inst_acq_compatibility.py
@@ -31,7 +31,7 @@ from aind_data_schema.components.devices import (
     PatchCord,
     ProbePort,
 )
-from aind_data_schema.components.calibrations import Calibration
+from aind_data_schema.components.measurements import Calibration
 from aind_data_schema.core.instrument import Instrument, Connection, ConnectionData, ConnectionDirection
 from aind_data_schema.core.acquisition import (
     Acquisition,
